requirements.txt
azure-storage-blob
azure-storage-file-share
azure-ai-formrecognizer
flask
python-dotenv
pytest
pytest-mock
requests
chardet 
tabulate 
python-barcode[images]

